Upwork-Copilot/requirements.txt
fastapi
uvicorn
pydantic
zhipuai
python-dotenv